--- a/hw10_program_optimization/BenchmarkGetDomainStat_new2.txt
+++ b/hw10_program_optimization/BenchmarkGetDomainStat_new2.txt
@@ -1,0 +1,16 @@
+goos: windows
+goarch: amd64
+pkg: github.com/diletech/hw-test/hw10_program_optimization
+cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
+BenchmarkGetDomainStat-8   	       4	 316777750 ns/op	29122318 B/op	  723197 allocs/op
+BenchmarkGetDomainStat-8   	       4	 316153950 ns/op	29122344 B/op	  723197 allocs/op
+BenchmarkGetDomainStat-8   	       4	 310334825 ns/op	29122396 B/op	  723197 allocs/op
+BenchmarkGetDomainStat-8   	       4	 309094750 ns/op	29122580 B/op	  723198 allocs/op
+BenchmarkGetDomainStat-8   	       4	 309951125 ns/op	29122232 B/op	  723197 allocs/op
+BenchmarkGetDomainStat-8   	       4	 317923450 ns/op	29122682 B/op	  723199 allocs/op
+BenchmarkGetDomainStat-8   	       4	 316795575 ns/op	29122992 B/op	  723201 allocs/op
+BenchmarkGetDomainStat-8   	       4	 321320600 ns/op	29122656 B/op	  723199 allocs/op
+BenchmarkGetDomainStat-8   	       4	 316395325 ns/op	29122564 B/op	  723199 allocs/op
+BenchmarkGetDomainStat-8   	       4	 311461775 ns/op	29122892 B/op	  723200 allocs/op
+PASS
+ok  	github.com/diletech/hw-test/hw10_program_optimization	25.439s