## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

---
```
goos: windows
goarch: amd64
pkg: github.com/diletech/hw-test/hw10_program_optimization
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                │ BenchmarkGetDomainStat_new2.txt │   BenchmarkGetDomainStat_old2.txt   │
                │             sec/op              │   sec/op     vs base                │
GetDomainStat-8                       316.3m ± 2%   615.5m ± 1%  +94.62% (p=0.000 n=10)

                │ BenchmarkGetDomainStat_new2.txt │     BenchmarkGetDomainStat_old2.txt     │
                │              B/op               │     B/op       vs base                  │
GetDomainStat-8                      27.77Mi ± 0%   308.32Mi ± 0%  +1010.11% (p=0.000 n=10)

                │ BenchmarkGetDomainStat_new2.txt │    BenchmarkGetDomainStat_old2.txt    │
                │            allocs/op            │  allocs/op    vs base                 │
GetDomainStat-8                       723.2k ± 0%   3045.4k ± 0%  +321.10% (p=0.000 n=10)

cpu: Intel(R) Core(TM) i7-9700KF CPU @ 3.60GHz
                │ BenchmarkGetDomainStat_new.txt │   BenchmarkGetDomainStat_old.txt    │
                │             sec/op             │   sec/op     vs base                │
GetDomainStat-8                      236.0m ± 1%   447.3m ± 2%  +89.54% (p=0.000 n=10)

                │ BenchmarkGetDomainStat_new.txt │     BenchmarkGetDomainStat_old.txt      │
                │              B/op              │     B/op       vs base                  │
GetDomainStat-8                     27.77Mi ± 0%   308.35Mi ± 0%  +1010.23% (p=0.000 n=10)

                │ BenchmarkGetDomainStat_new.txt │    BenchmarkGetDomainStat_old.txt     │
                │           allocs/op            │  allocs/op    vs base                 │
GetDomainStat-8                      723.2k ± 0%   3045.4k ± 0%  +321.10% (p=0.000 n=10)
```